Cleanup:
- Moved around book from `examples` to `candle-book` proper (overlapping
  the book and the lib structures)
