+++ candle-book/src/training/mnist.md
@@ -3,7 +3,7 @@
 So we now have downloaded the MNIST parquet files, let's put them in a simple struct.
 
 ```rust,ignore
-{{#include ../../../candle-examples/src/lib.rs:book_training_3}}
+{{#include ../lib.rs:book_training_3}}
 ```
 
 The parsing of the file and putting it into single tensors requires the dataset to fit the entire memory.
